# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Contacts Application",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "RESTful API for managing personal contacts. Users can register, log in, and perform CRUD operations on their contacts. The API uses token-based authentication via access and refresh tokens stored in sessions. Built with Node.js, Express, and MongoDB."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Local server"
    },
    {
      "url": "https://contacts-app-l60g.onrender.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
    "/contacts": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get list of contacts",
        "operationId": "getContacts",
        "description": "Retrieves a paginated, sorted, and optionally filtered list of contacts belonging to the authenticated user.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "example": 1,
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "perPage",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "example": 10,
            "description": "Number of contacts per page"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string"
            },
            "example": "name",
            "description": "Field to sort by"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "example": "asc",
            "description": "Sort direction"
          },
          {
            "in": "query",
            "name": "isFavourite",
            "schema": {
              "type": "boolean"
            },
            "example": true,
            "description": "Filter contacts by their status"
          },
          {
            "in": "query",
            "name": "contactType",
            "schema": {
              "type": "string",
              "enum": [
                "work",
                "home",
                "personal"
              ]
            },
            "example": "home",
            "description": "Filter contacts by type"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found contacts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found contacts!"
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "data",
                        "page",
                        "perPage",
                        "totalItems",
                        "totalPages",
                        "hasNextPage",
                        "hasPreviousPage"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/contacts"
                          }
                        },
                        "page": {
                          "type": "number",
                          "example": 1
                        },
                        "perPage": {
                          "type": "number",
                          "example": 10
                        },
                        "totalItems": {
                          "type": "number",
                          "example": 20
                        },
                        "totalPages": {
                          "type": "number",
                          "example": 2
                        },
                        "hasNextPage": {
                          "type": "boolean",
                          "example": true
                        },
                        "hasPreviousPage": {
                          "type": "boolean",
                          "example": false
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "tags": [
          "Contacts"
        ],
        "summary": "Add a new contact",
        "operationId": "addContact",
        "description": "Creates a new contact for the authenticated user. A photo can be uploaded",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "phoneNumber",
                  "contactType",
                  "isFavourite"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Contact's  name",
                    "example": "Alice Johnson"
                  },
                  "email": {
                    "type": "string",
                    "description": "Contact's  email",
                    "format": "email",
                    "example": "alice.johnson@example.com"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "description": "Contact's phone number",
                    "example": "+380000000001"
                  },
                  "contactType": {
                    "type": "string",
                    "description": "Contact's type",
                    "enum": [
                      "work",
                      "home",
                      "personal"
                    ],
                    "example": "home"
                  },
                  "isFavourite": {
                    "type": "boolean",
                    "description": "Contact's priority",
                    "example": false
                  },
                  "photo": {
                    "type": "string",
                    "description": "Contact's photo",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contact successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Succesfully created a contact"
                    },
                    "data": {
                      "$ref": "#/components/schemas/contacts"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "$ref": "#/components/responses/400"
          },
          "401": {
            "description": "Unauthorized",
            "$ref": "#/components/responses/401"
          },
          "500": {
            "description": "Internal Server Error",
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/contacts/{contactId}": {
      "get": {
        "tags": [
          "Contacts"
        ],
        "summary": "Get contact by id",
        "operationId": "getContactById",
        "description": "Retrieves a single contact by its unique ID for the authenticated user.\n    Only the owner of the contact can access this resource.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The unique identifier of the contact to retrieve",
              "example": "67f925cbb795599b632edfe0"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully get contact by id",
            "content": {
              "aplication/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found contact with id ${contactId}!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/contacts"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "$ref": "#/components/responses/400"
          },
          "401": {
            "description": "Unauthorized",
            "$ref": "#/components/responses/401"
          },
          "404": {
            "description": "Contact with the specified ID was not found",
            "$ref": "#/components/responses/404"
          },
          "500": {
            "description": "Internal Server Error",
            "$ref": "#/components/responses/500"
          }
        }
      },
      "patch": {
        "tags": [
          "Contacts"
        ],
        "summary": "Update a contact (partial)",
        "operationId": "patchContact",
        "description": "Updates one or more fields of an existing contact.\nOnly the contact's owner can perform this action.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "661a2cd3c991f1a9fba1dce1",
            "description": "The ID of the contact to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Contact's name",
                    "example": "Jane Doe"
                  },
                  "email": {
                    "type": "string",
                    "description": "Contact's email",
                    "format": "email",
                    "example": "jane.doe@example.com"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "description": "Contact's phone number",
                    "example": "+380000000001"
                  },
                  "contactType": {
                    "type": "string",
                    "description": "Contact's type",
                    "enum": [
                      "work",
                      "home",
                      "personal"
                    ],
                    "example": "work"
                  },
                  "isFavourite": {
                    "type": "boolean",
                    "description": "Contact's prio",
                    "example": true
                  },
                  "photo": {
                    "type": "string",
                    "description": "Contact's photo",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully patched a contact!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/contacts"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "$ref": "#/components/responses/401"
          },
          "404": {
            "description": "Contact with the specified ID was not found",
            "$ref": "#/components/responses/404"
          },
          "500": {
            "description": "Internal Server Error",
            "$ref": "#/components/responses/500"
          }
        }
      },
      "put": {
        "tags": [
          "Contacts"
        ],
        "summary": "Upsert a contact (create or update)",
        "operationId": "upsertContact",
        "description": "Creates a new contact or updates an existing one based on the provided ID.\nThis is an \"upsert\" operation: if the contact doesn't exist, it will be created.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "661a2cd3c991f1a9fba1dce1",
            "description": "The ID of the contact to create or update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "phoneNumber",
                  "contactType",
                  "isFavourite"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Contact's  name",
                    "example": "Alice Johnson"
                  },
                  "email": {
                    "type": "string",
                    "description": "Contact's  email",
                    "format": "email",
                    "example": "alice.johnson@example.com"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "description": "Contact's phone number",
                    "example": "+380000000001"
                  },
                  "contactType": {
                    "type": "string",
                    "description": "Contact's type",
                    "enum": [
                      "work",
                      "home",
                      "personal"
                    ],
                    "example": "home"
                  },
                  "isFavourite": {
                    "type": "boolean",
                    "description": "Contact's priority",
                    "example": false
                  },
                  "photo": {
                    "type": "string",
                    "description": "Contact's photo",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully patched a contact!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/contacts"
                    }
                  }
                }
              }
            }
          },
          "201": {
            "description": "Contact created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully patched a contact!"
                    },
                    "data": {
                      "$ref": "#/components/schemas/contacts"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "$ref": "#/components/responses/401"
          },
          "500": {
            "description": "Internal Server Error",
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "tags": [
          "Contacts"
        ],
        "summary": "Delete a contact",
        "operationId": "deleteContact",
        "description": "Deletes a contact by its ID. Only the owner of the contact can delete it.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "contactId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "661a2cd3c991f1a9fba1dce1",
            "description": "The ID of the contact to delete"
          }
        ],
        "responses": {
          "204": {
            "description": "Contact successfully deleted (no content)"
          },
          "401": {
            "description": "Unauthorized",
            "$ref": "#/components/responses/401"
          },
          "404": {
            "description": "Contact with the specified ID was not found",
            "$ref": "#/components/responses/404"
          },
          "500": {
            "description": "Internal Server Error",
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "contacts": {
        "type": "object",
        "required": [
          "_id",
          "userId",
          "name",
          "phone",
          "contactType",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Contact's id",
            "example": "661a2cd3c991f1a9fba1dce1"
          },
          "userId": {
            "type": "string",
            "description": "Contact's crestor  id",
            "example": "661a2cbfc991f1a9fba1dcdb"
          },
          "name": {
            "type": "string",
            "description": "Contact's name",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "description": "Contact's email",
            "example": "john.doe@example.com"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Contact's phone number",
            "example": "+380000000001"
          },
          "contactType": {
            "type": "string",
            "description": "Contact's type",
            "enum": [
              "work",
              "home",
              "personal"
            ],
            "example": "personal"
          },
          "isFavourite": {
            "type": "boolean",
            "description": "Contact's priority",
            "example": false
          },
          "photo": {
            "type": "string",
            "description": "Contact's photo",
            "example": "https://res.cloudinary.com/.../photo.jpg"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-04-01T10:22:30.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-04-01T10:22:30.000Z"
          }
        }
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "Error"
                },
                "data": {
                  "type": "string",
                  "example": "Request body could not be read properly"
                }
              }
            }
          }
        }
      },
      "401": {
        "description": "Unauthorized – Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "Unauthorized Error"
                },
                "data": {
                  "type": "string",
                  "example": "Access token is expired"
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Not Found – The requested resource does not exist",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "Not Found Error"
                },
                "data": {
                  "type": "string",
                  "example": "Contact not found"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    }
  }
}